\name{anominate-package}
\alias{anominate-package}
\docType{package}
\title{alpha-NOMINATE}
\description{Fits ideal point model described in Carroll, Lewis, Lo, Poole and Rosenthal, "The Structure of Utility in Models of Spatial Voting," American Journal of Political Science 57(4): 1008--1028.}
\details{
\tabular{ll}{
Package: \tab anominate\cr
Type: \tab Package\cr
Version: \tab 0.6\cr
Date: \tab 2017-10-07\cr
License: \tab GPL-2 \cr
}}
\author{Royce Carroll <rcarroll@rice.edu>, Christopher Hare <cdhare@ucdavis.edu>, Jeffrey B. Lewis <jblewis@ucla.edu>, James Lo <lo@uni-mannheim.de>, Keith T. Poole <ktpoole@uga.edu>, and Howard Rosenthal <hl31@nyu.edu>}
\references{Carroll, Lewis, Lo, Poole and Rosenthal, "The Structure of Utility in Models of Spatial Voting," American Journal of Political Science 57(4): 1008--1028.}
\keyword{anominate}
\seealso{wnominate}


